+++ linux/include/flutter_device_info_plus/flutter_device_info_plus_plugin.h
@@ -1,0 +1,9 @@
+#ifndef FLUTTER_DEVICE_INFO_PLUS_PLUGIN_H_
+#define FLUTTER_DEVICE_INFO_PLUS_PLUGIN_H_
+
+#include <flutter/plugin_registrar_linux.h>
+
+void FlutterDeviceInfoPlusPluginRegisterWithRegistrar(
+    flutter::PluginRegistrarLinux *registrar);
+
+#endif  // FLUTTER_DEVICE_INFO_PLUS_PLUGIN_H_